+++ groovy-android-gradle-plugin/src/test/groovy/groovyx/functional/SourceSetSpec.groovy
@@ -24,6 +24,7 @@ import java.lang.Void as Should
 import static groovyx.internal.TestProperties.allTests
 import static groovyx.internal.TestProperties.androidPluginVersion
 import static groovyx.internal.TestProperties.compileSdkVersion
+import static groovyx.internal.TestProperties.groovyVersion
 
 @IgnoreIf({ !allTests })
 class SourceSetSpec extends AndroidFunctionalSpec {
@@ -74,8 +75,8 @@ class SourceSetSpec extends AndroidFunctionalSpec {
         }
 
         compileOptions {
-          sourceCompatibility '1.7'
-          targetCompatibility '1.7'
+          sourceCompatibility '1.8'
+          targetCompatibility '1.8'
         }
 
          buildTypes {
@@ -100,7 +101,7 @@ class SourceSetSpec extends AndroidFunctionalSpec {
       }
 
       dependencies {
-        implementation 'org.codehaus.groovy:groovy:2.4.16:grooid'
+        implementation 'org.codehaus.groovy:groovy:$groovyVersion:grooid'
       }
     """
 
@@ -230,8 +231,8 @@ class SourceSetSpec extends AndroidFunctionalSpec {
         }
 
         compileOptions {
-          sourceCompatibility '1.7'
-          targetCompatibility '1.7'
+          sourceCompatibility '1.8'
+          targetCompatibility '1.8'
         }
 
         buildTypes {
@@ -256,7 +257,7 @@ class SourceSetSpec extends AndroidFunctionalSpec {
       }
 
       dependencies {
-        implementation 'org.codehaus.groovy:groovy:2.4.16:grooid'
+        implementation 'org.codehaus.groovy:groovy:$groovyVersion:grooid'
       }
     """